+++ artifact-restore.sh
@@ -1,9 +1,9 @@
 #!/usr/bin/env bash
 
 ################################################################################
-# Bash script that copies lastest weekly backup from an Artifactory server on
-# Linux to a Windows network share. Including removing oldest weekly backup
-# from network share.
+# Bash script that copies lastest weekly backup from a Windows network share
+# an Artifactory server on Linux. In case of desaster recovery this need to
+# happen before backup import using Artifactory web interface.
 
 set -u # Exit script when using an uninitialised variable
 set -e # Exit script when a statement returns a non-true value
@@ -19,70 +19,37 @@ set -e # Exit script when a statement returns a non-true value
 
 SCRIPT_DIR=$(dirname "${BASH_SOURCE[0]}")
 CFG_FILE="live-cfg.sh"
-DIRECTION="to"
+DIRECTION="from"
 
 
 ################################################################################
 # Functions
 
-check_idle_artifactory()
+restore_newest_weekly()
 {
-  local unfinished
-  unfinished=$(find "${FULL_LOCAL_DIR}" -maxdepth 1 -type d -regextype sed -regex "${TMP_DIR_REGEX}")
-  if [ -n "${unfinished}" ]; then
-    echo "New weekly backup is in progress and it isn't safe to proceed."
-    close_network
-    exit 1
-  fi
-}
-
-backup_newest_weekly()
-{
-  local local_newest remote_newest tarfile
+  local local_newest remote_newest
   cd "${FULL_REMOTE_DIR}"
   remote_newest=$(find . -maxdepth 1 -type f -regextype sed -regex "${BACKUP_FILE_REGEX}" | sort -n -r | head -1)
   if [ -n "${remote_newest}" ] && [ ! -s "${remote_newest}" ]; then
-    echo "Remote tarball is zero bytes, removing it."
-    rm -f "${remote_newest}"
+    echo "Remote tarball is zero bytes, unable to use it."
+    exit 1
   fi
   cd -
   cd "${FULL_LOCAL_DIR}"
   local_newest=$(find . -maxdepth 1 -type d -regextype sed -regex "${BACKUP_DIR_REGEX}" | sort -n -r | head -1)
   if [ "${remote_newest}" = "${local_newest}.tar" ]; then
-    echo "Up-to-date backup"
-    return
+    echo "Local backup is up-to-date."
+    exit 0
   fi
-  tarfile="${FULL_REMOTE_DIR}/${local_newest}.tar"
-  echo "Backup needs to be transferred."
-  echo "  local: ${local_newest}"
-  echo "  remote: ${tarfile}"
-  tar -cf "${tarfile}" "${local_newest}"
-  cd -
+  echo "Backup needs to be restored."
+  echo "  remote: ${FULL_REMOTE_DIR}/${remote_newest}"
+  echo "  local: ${FULL_LOCAL_DIR}"
+  tar -xf "${FULL_REMOTE_DIR}/${remote_newest}"
 }
-
-cleanup_network_backups()
-{
-  remote_count=$(find "${FULL_REMOTE_DIR}" -maxdepth 1 -type f -regextype sed -regex "${BACKUP_FILE_REGEX}" | wc -l)
-  if [ -n "${remote_count}" ] && [[ "${remote_count}" -gt ${BACKUP_COUNT} ]]; then
-    remote_oldest=$(find "${FULL_REMOTE_DIR}" -maxdepth 1 -type f -regextype sed -regex "${BACKUP_FILE_REGEX}" | sort -n | head -1)
-    echo "Removing oldest tarball backup - ${remote_oldest}"
-    rm -f "${remote_oldest}"
-  fi
-}
-
-close_network()
-{
-  local mounted
-  mounted=$(mount | grep "${MOUNT_POINT}" || true)
-  if [ -n "${mounted}" ]; then
-    echo "Unmounting remote backup machine!"
-    umount "${MOUNT_POINT}"
-  fi
-}
-
 
 ################################################################################
 # Main
+
 
 # shellcheck source=common.sh
 source "${SCRIPT_DIR}/common.sh"
@@ -90,10 +57,10 @@ TEST=0
 while [[ "$#" -gt 0 ]]; do
   case $1 in
   "-t" | "--test")
-    if [[ "$#" -eq 2 ]]; then
+    if [[ "$#" -gt 1 ]]; then
       test_cfg="${SCRIPT_DIR}/test/test-$2-cfg.sh"
       if [ -s "${test_cfg}" ]; then
-        # shellcheck source=test/test-1-cfg.sh
+        # shellcheck source=test/test-restore-cfg.sh
         source "${test_cfg}"
         FULL_REMOTE_DIR="${MOUNT_POINT}/${REMOTE_DIR}"
         TEST=1
@@ -122,7 +89,6 @@ while [[ "$#" -gt 0 ]]; do
   esac
 done
 
-
 if [ ${TEST} -eq 0 ]; then
   if [ -s "${SCRIPT_DIR}/${CFG_FILE}" ]; then
     # shellcheck source=template-cfg.sh
@@ -135,9 +101,6 @@ if [ ${TEST} -eq 0 ]; then
   check_config
   check_mount_point
 fi
-check_idle_artifactory
-backup_newest_weekly
-cleanup_network_backups
-close_network
+restore_newest_weekly
 
 exit 0