+++ CHANGELOG.md
@@ -5,7 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - 2019-??-?? [Unrelease]
+## [0.2.0] - 2019-xx-xx [Unrelease]
+
+### Added
+
+- Restore script is idempotent.
+- Added configration file parameter BACKUP_COUNT (was in artifact-backup.sh)
+- Added configration file parameter CIFS_VERSION (was in artifact-backup.sh)
+
+### Changed
+
+- Renamed configuration file parameter FULL_LOCAL_DIR (before: FULL_SRC_DIR)
+- Renamed configuration file parameter REMOTE_DIR (before: DEST_DIR)
+
+## [0.1.0] - 2019-02-14
 
 ### Added